requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pyswisseph==2.10.3.2
pytz==2024.2
Werkzeug==3.0.4
